:recycle: REFACTOR: 서비스 가동률 컨트롤러 이동
