+++ slash/src/main/java/project/slash/statistics/controller/StatisticsController.java
@@ -29,22 +29,6 @@ import project.slash.statistics.service.StatisticsService;
 public class StatisticsController {
 	private final StatisticsService statisticsService;
 
-	@GetMapping("/common/statistics")
-	public BaseResponse<?> getServiceUptimeStatistics(
-		@RequestParam("date") @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME) LocalDate date,
-		@RequestParam("evaluationItemId") long evaluationItemId) {
-		List<MonthlyStatisticsDto> statisticsList = statisticsService.calculateMonthlyStats(date, evaluationItemId);
-		return BaseResponse.ok(statisticsList);
-	}
-
-	@PostMapping("/contract-manager/statistics")
-	public BaseResponse<?> addStatistics(@RequestBody RequestStatisticsDto requestStatisticsDto) {
-		statisticsService.createMonthlyStats(requestStatisticsDto);
-
-		return BaseResponse.ok();
-	}
-
-
 	/**
 	 * 월간 지표 조회하는 메서드입니다.
 	 *
